#Quote and Invoice Changes

+++ Objects/SeedData/TaxCodeSeedModel.cs
@@ -5,12 +5,12 @@ namespace Vanigam.CRM.Objects.SeedData
         public string Code { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
         public string TaxType { get; set; } = string.Empty;
-        public decimal TaxRate { get; set; }
-        public decimal CGSTRate { get; set; }
-        public decimal SGSTRate { get; set; }
-        public decimal IGSTRate { get; set; }
-        public decimal UTGSTRate { get; set; }
-        public decimal CessRate { get; set; }
+        public double TaxRate { get; set; }
+        public double CGSTRate { get; set; }
+        public double SGSTRate { get; set; }
+        public double IGSTRate { get; set; }
+        public double UTGSTRate { get; set; }
+        public double CessRate { get; set; }
         public bool IsActive { get; set; }
         public bool IsCompoundTax { get; set; }
     }
